backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
branca==0.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.13.0
django-extensions==3.2.1
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.26.1
drf-yasg==1.21.5
folium==0.14.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
googlemaps==4.10.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-google-places==1.4.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
ratelim==0.1.6
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
